+++ index.ts
@@ -161,6 +161,11 @@ const tslintPlugin = <TslintPlugin> function(pluginOptions?: PluginOptions) {
 };
 
 tslintPlugin.report = function(options?: ReportOptions) {
+    // Notify the user that the old interface is used, this can be removed at some point
+    if (isString(options)) {
+        throw new Error("Deprecated interface used! See 6.0.0 changelog " +
+            "https://github.com/panuhorsmalahti/gulp-tslint/blob/master/CHANGELOG.md");
+    }
 
     // Default options
     if (!options) {